Done
I managed to get all my information displayed and aligned where I want it. I was going to follow along with his version and then compare the two, but he never posted it so mine is the only one. I will make it look nicer and explain the code on the app in the future but for now this is done.